+++ kbl/kbl-v24/src/main/java/com/foursoft/harness/kbl/v24/HasPlacement.java
@@ -1,6 +1,6 @@
 /*-
  * ========================LICENSE_START=================================
- * KBL Common
+ * KBL 2.5
  * %%
  * Copyright (C) 2020 - 2023 4Soft GmbH
  * %%
@@ -23,12 +23,12 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.kbl.common;
+package com.foursoft.harness.kbl.v24;
 
-public interface HasPlacement<X> {
+public interface HasPlacement {
 
-    X getPlacement();
+    KblTransformation getPlacement();
 
-    void setPlacement(X value);
+    void setPlacement(KblTransformation value);
 
 }
