+++ kbl/kbl-v24/src/main/resources/kbl24/KBL24_SR1-ext.xjb
@@ -94,7 +94,7 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Placement']]">
-        <inheritance:implements>com.foursoft.harness.kbl.common.HasPlacement&lt;com.foursoft.harness.kbl.v24.KblTransformation&gt;</inheritance:implements>
+        <inheritance:implements>com.foursoft.harness.kbl.v24.HasPlacement</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Instruction_type']]">
