+++ kbl/kbl-v25/src/main/resources/kbl25/KBL25_SR1-ext.xjb
@@ -97,7 +97,7 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Placement']]">
-        <inheritance:implements>com.foursoft.harness.kbl.common.HasPlacement&lt;com.foursoft.harness.kbl.v25.KblTransformation&gt;</inheritance:implements>
+        <inheritance:implements>com.foursoft.harness.kbl.v25.HasPlacement</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Instruction_type']]">
